beta 2 commit

+++ app/View/Settings/admin_index.ctp
@@ -1,12 +1,39 @@
-<h1 class="pull-left">
-    Settings Categories
-</h1>
-<?= $this->Html->link('Add Settings Category <i class="icon icon-plus icon-white"></i>', array('action' => 'add'), array(
-    'class' => 'btn btn-info pull-right', 
-    'style' => 'margin-bottom:10px',
-    'escape' => false
-)) ?>
+<?php $this->Html->addCrumb('Admin', '/admin') ?>
+<?php $this->Html->addCrumb('Settings', null) ?>
+
+<div class="pull-left">
+    <h1>Setting Categories<?php if (!empty($this->params->named['trash'])): ?> - Trash<?php endif ?></h1>
+</div>
+<div class="btn-group" style="float:right;margin-bottom:10px">
+  <a class="btn dropdown-toggle" data-toggle="dropdown">
+    View <i class="icon-picture"></i>
+    <span class="caret"></span>
+  </a>
+  <ul class="dropdown-menu view">
+    <li>
+        <?= $this->Html->link('<i class="icon-ok"></i> Active', array(
+            'admin' => true, 
+            'action' => 'index'
+        ), array('escape' => false)) ?>
+    </li>
+    <li>
+        <?= $this->Html->link('<i class="icon-trash"></i> Trash', array(
+            'admin' => true, 
+            'action' => 'index', 
+            'trash' => 1
+        ), array('escape' => false)) ?>
+    </li>
+  </ul>
+</div>
 <div class="clearfix"></div>
+
+<?php if ($this->Admin->hasPermission($permissions['related']['settings']['admin_add'])): ?>
+    <?= $this->Html->link('Add Settings Category <i class="icon icon-plus icon-white"></i>', array('action' => 'add'), array(
+        'class' => 'btn btn-info pull-right', 
+        'style' => 'margin-bottom:10px',
+        'escape' => false
+    )) ?>
+<?php endif ?>
 
 <?php if (empty($this->request->data)): ?>
     <div class="well">
@@ -26,8 +53,10 @@
             <tbody>
                 <tr>
                     <td>
-                        <?= $this->Html->link($data['Setting']['title'], array(
-                            'action' => 'edit', $data['Setting']['id'])); ?>
+                        <?php if ($this->Admin->hasPermission($permissions['related']['settings']['admin_edit'], $data['Setting']['id'])): ?>
+                            <?= $this->Html->link($data['Setting']['title'], array(
+                            'action' => 'edit', $data['Setting']['id'])) ?>
+                        <?php endif ?>
                     </td>
                     <td>
                         <?= $this->Admin->time($data['Setting']['created']) ?>
@@ -40,32 +69,40 @@
                             </a>
                             <ul class="dropdown-menu">
                                 <?php if (empty($this->params->named['trash'])): ?>
-                                    <li>
-                                        <?= $this->Admin->edit(
-                                            $data['Setting']['id']
-                                        ) ?>
-                                    </li>
-                                    <li>
-                                        <?= $this->Admin->delete(
-                                            $data['Setting']['id'],
-                                            $data['Setting']['title'],
-                                            'setting category'
-                                        ) ?>
-                                    </li>
+                                    <?php if ($this->Admin->hasPermission($permissions['related']['settings']['admin_edit'], $data['Setting']['id'])): ?>
+                                        <li>
+                                            <?= $this->Admin->edit(
+                                                $data['Setting']['id']
+                                            ) ?>
+                                        </li>
+                                    <?php endif ?>
+                                    <?php if ($this->Admin->hasPermission($permissions['related']['settings']['admin_delete'], $data['Setting']['id'])): ?>
+                                        <li>
+                                            <?= $this->Admin->delete(
+                                                $data['Setting']['id'],
+                                                $data['Setting']['title'],
+                                                'setting category'
+                                            ) ?>
+                                        </li>
+                                    <?php endif ?>
                                 <?php else: ?>
-                                    <li>
-                                        <?= $this->Admin->restore(
-                                            $data['Setting']['id'],
-                                            $data['Setting']['title']
-                                        ) ?>
-                                    </li>  
-                                    <li>
-                                        <?= $this->Admin->delete_perm(
-                                            $data['Setting']['id'],
-                                            $data['Setting']['title'],
-                                            'setting category'
-                                        ) ?>
-                                    </li>     
+                                    <?php if ($this->Admin->hasPermission($permissions['related']['settings']['admin_restore'], $data['Setting']['id'])): ?>
+                                        <li>
+                                            <?= $this->Admin->restore(
+                                                $data['Setting']['id'],
+                                                $data['Setting']['title']
+                                            ) ?>
+                                        </li>
+                                    <?php endif ?>
+                                    <?php if ($this->Admin->hasPermission($permissions['related']['settings']['admin_delete'], $data['Setting']['id'])): ?>
+                                        <li>
+                                            <?= $this->Admin->delete_perm(
+                                                $data['Setting']['id'],
+                                                $data['Setting']['title'],
+                                                'setting category'
+                                            ) ?>
+                                        </li>  
+                                    <?php endif ?>   
                                 <?php endif ?>
                             </ul>
                         </div>
